Quality of life changes
created npm run dev so that there is no need to run 2 terminals (1 for frontend and backend) and also no need to start a venv everytime I start working on the project.

Also created a manifest.json to remove the console error shown in browser inpsect. This will be useful as I can update it as i go rather than all at once.